+++ CHANGELOG.md
@@ -7,6 +7,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 * [FuiPopover] - Removed `react-tiny-popover` and replace with `floating-ui`
+* [FuiOptionGroup] - New Component 🎉
 
 ## [0.0.2] - 2023-12-21
 Initial Release 🎉